Add break from while loop if the order is found and complete responses for handling order

+++ checkout/webhook_handler.py
@@ -58,45 +58,49 @@ class StripeWH_Handler:
                     stripe_pid=pid,
                 )
                 order_exists = True
-                return HttpResponse(
-                    content=f'Webhook received: {event['type']}',
-                    status=200)
+                break
             except Order.DoesNotExist:
                 attempt += 1
                 time.sleep(1)
-                try:
-                    order = Order.objects.create(
-                            first_name=shipping_details.name.split(' ')[0],
-                            last_name=shipping_details.name.split(' ')[1],
-                            email=billing_details.email,
-                            mobile_number=shipping_details.phone,
-                            country=shipping_details.address.country,
-                            postcode=shipping_details.address.postal_code,
-                            town_or_city=shipping_details.address.city,
-                            address_line1=shipping_details.address.line1,
-                            address_line2=shipping_details.address.line2,
-                            county=shipping_details.address.state,
-                            original_bag=bag,
-                            stripe_pid=pid,
+        if order_exists:
+            return HttpResponse(
+                    content=f'Webhook received: {event['type']} | SUCCESS: Verified order already in database',
+                    status=200)
+        else:
+            order = None
+            try:
+                order = Order.objects.create(
+                        first_name=shipping_details.name.split(' ')[0],
+                        last_name=shipping_details.name.split(' ')[1],
+                        email=billing_details.email,
+                        mobile_number=shipping_details.phone,
+                        country=shipping_details.address.country,
+                        postcode=shipping_details.address.postal_code,
+                        town_or_city=shipping_details.address.city,
+                        address_line1=shipping_details.address.line1,
+                        address_line2=shipping_details.address.line2,
+                        county=shipping_details.address.state,
+                        original_bag=bag,
+                        stripe_pid=pid,
+                    )
+                for item_id, item_data in json.loads(bag).items():
+                    product = Product.objects.get(id=item_id)
+                    if isinstance(item_data, int):
+                        order_line_item = OrderLineItem(
+                            order=order,
+                            product=product,
+                            quantity=item_data,
                         )
-                    for item_id, item_data in json.loads(bag).items():
-                        product = Product.objects.get(id=item_id)
-                        if isinstance(item_data, int):
-                            order_line_item = OrderLineItem(
-                                order=order,
-                                product=product,
-                                quantity=item_data,
-                            )
-                            order_line_item.save()
-                except Exception as e:
-                    if order:
-                        order.delete()
-                    return HttpResponse(
-                        content=f'Webhook received: {event["type"]} | ERROR: {e}',
-                        status=500)       
+                        order_line_item.save()
+            except Exception as e:
+                if order:
+                    order.delete()
+                return HttpResponse(
+                    content=f'Webhook received: {event["type"]} | ERROR: {e}',
+                    status=500)       
             
         return HttpResponse(
-                content=f'Webhook received: {event["type"]}',
+                content=f'Webhook received: {event["type"]} | SUCCESS: Created order in webhook',
                 status=200)
         
     def handle_payment_intent_payment_failed(self, event):
